[feat] 네트워크 모델 정리
content
- CardDisplayable 프로토콜 선언
- Todo, NewCard 모델 프로토콜 채택
- 프로토콜을 이용해 모델의 데이터 추출

[#25]
