+++ iOS/To-Do_List/To-Do_List/Model/NetworkResult.swift
@@ -1,0 +1,18 @@
+//
+//  NetworkResult.swift
+//  To-Do_List
+//
+//  Created by 박진섭 on 2022/04/14.
+//
+
+import Foundation
+
+struct NetworkResult:Codable {
+    let response:Board
+}
+
+extension NetworkResult:Equatable {
+    static func == (lhs: NetworkResult, rhs: NetworkResult) -> Bool {
+        lhs.response.completedItems == rhs.response.completedItems && lhs.response.progressingItems == rhs.response.progressingItems && lhs.response.todoItems == rhs.response.todoItems
+    }
+}
