+++ iOS/To-Do_List/To-Do_List/Model/Todo.swift
@@ -7,26 +7,15 @@
 
 import Foundation
 
-//TODO: 타입명 변경
-struct NetworkResult:Codable {
-    let response:Board
+protocol CardDisplayable {
+    var id: Int? { get }
+    var title: String { get }
+    var content: String { get }
 }
 
-struct Board:Codable {
-    let todoItems:[Todo]
-    let progressingItems:[Todo]
-    let completedItems:[Todo]
-}
-
-struct Todo:Codable,Equatable {
-    let id:Int
+struct Todo: CardDisplayable, Codable, Equatable {
+    let id:Int?
     let title:String
     let content:String
     let createdAt:String
-}
-
-extension NetworkResult:Equatable {
-    static func == (lhs: NetworkResult, rhs: NetworkResult) -> Bool {
-        lhs.response.completedItems == rhs.response.completedItems && lhs.response.progressingItems == rhs.response.progressingItems && lhs.response.todoItems == rhs.response.todoItems
-    }
 }
